Only show progressbar while loading
- The progress bar was showing up while swiping down to refresh and so
causing to see a bit of jank where the screen would refresh and you see
the progressbar as well. A bit redundant. This shows it only on boot and
gives the user some kind of indication if they have a slow(er) connection.

- Also move 'swipe refresh layout' down to onPostExecute so you can enter
the app without a internet connection. With it in onPreExecute, you can't
enter the app and it just looks bad. It looks like there's something wrong
wtih the app and no indication that is because of internet connection.

Change-Id: I991288f91c4fd8d2de80eeb8b384fe0b65ec10d1